requirements.txt
clickhouse-driver==0.2.3
mysql==0.0.3
mysqlclient==2.0.3
numpy==1.22.3
pandas==1.4.2
PyYAML==6.0
SQLAlchemy==1.4.36